chore: optimize some things
