+++ src/packages/interactive/animation/src/bytedance-microapp/index.ts
@@ -1,7 +1,11 @@
 import { Animation, AnimationOptions } from '../types';
+import { CONTAINER_NAME } from '@utils/constant';
+import { normalizeOptions } from '../common';
 
 export function createAnimation(options?: AnimationOptions): Animation {
-  return tt.createAnimation(options);
+  return tt.createAnimation(
+    normalizeOptions(options, CONTAINER_NAME.BYTE),
+  );
 }
 
 export default {
