+++ src/packages/interactive/animation/src/web/index.ts
@@ -1,17 +1,10 @@
 import { Animation, AnimationOptions } from '../types';
 import AnimationImpl from './AnimationImpl';
-import { applyAnimation } from '../common';
 
-function createAnimation(options?: AnimationOptions): Animation {
+export function createAnimation(options?: AnimationOptions): Animation {
   return new AnimationImpl(options);
 }
 
-export {
-  createAnimation,
-  applyAnimation,
-};
-
 export default {
   createAnimation,
-  applyAnimation,
 };
